Feat(Problem): 문제 세트 생성 UI 완성
- 문제 세트 생성 UI 완성
- 기능은 추후 API 연동 시 추가 예정
- Header UI 변경 하였음
- 전체 폰트 Pretendard-Regular로 임시 결정

Jira issue: SWM-18
See also: SWM-42
