--- jabiseo/app/globals.css
@@ -1,10 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-}
-
-button {
-  padding: 0px;
-  border: none;
-  background-color: white;
-}
