+++ jabiseo/package.json
@@ -16,9 +16,10 @@
     "@mui/material": "^5.15.19",
     "@mui/material-nextjs": "^5.15.11",
     "@react-oauth/google": "^0.12.1",
-    "next": "14.2.3",
-    "react": "^18",
-    "react-dom": "^18",
+    "eslint-config-next": "^15.0.0-rc.0",
+    "next": "^15.0.0-rc.0",
+    "react": "^19.0.0-rc-dfd30974ab-20240613",
+    "react-dom": "^19.0.0-rc-dfd30974ab-20240613",
     "react-icons": "^5.2.1",
     "react-kakao-login": "^2.1.1",
     "react-markdown": "^9.0.1",
