+++ jabiseo/src/app/page.tsx
@@ -1,0 +1,10 @@
+import Appbar from "../components/Appbar";
+
+export default function Home() {
+  return (
+    <>
+      <Appbar />
+      <h1>awd</h1>
+    </>
+  );
+}
