+++ jabiseo/src/app/test/page.tsx
@@ -15,8 +15,8 @@ $$
   `;
   const mark2 = "![img](https://storage.googleapis.com/machuda/hwp/2023-04-19T15:33:32.334471)";
   const mark3 = `
-다음 설명에 해당하는 시스템으로 옳은 것은?<br> <br><img width="300" src="https://storage.googleapis.com/machuda/hwp/2023-04-19T11:47:26.596099500" /> asda<br>
-awdawdawd
+다음 두 릴레이션에서 외래키로 사용된 것은? (단 밑줄 친 속성은 기본키이다.)<br>     <br>![20년_1,2회차 42번](https://storage.googleapis.com/machuda/hwp/2023-04-19T15:33:33.183993)<br>
+
   `;
   return (
     <Box
